Create a variadic module and move code around accordingly
